+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `ndh.tests.TestNDHLinks`
+
 ## [v6.1.0] - 2023-04-03
 
 - improve `Links` with getters for create, delete, detail, list & update urls